requirements.txt
flask==2.3.3
flask-cors==4.0.0
gunicorn==20.1.0
python-dotenv==1.0.1
openai>=1.31.0
numpy==1.26.4
scipy==1.13.0
markdown==3.6
requests==2.31.0
beautifulsoup4==4.12.3
PyPDF2==3.0.1
markdownify==0.11.6
